Move `hom_alt_depletion_fix` to a utils file for use in other scripts
